build: finish renaming ihe gw outbound result functions
Ref. metriport/metriport-internal#1350
