+++ packages/api/src/external/carequality/outbound-document-query-resp.ts
@@ -1,4 +1,4 @@
-import { OutboundDocumentQueryResp as OutboundDocumentQueryRespCore } from "@metriport/core/external/carequality/ihe-result";
+import { OutboundDocumentQueryResp as OutboundDocumentQueryRespCore } from "@metriport/core/external/carequality/ihe-gateway/ihe-result";
 import { BaseDomainCreate } from "@metriport/core/domain/base-domain";
 
 export interface OutboundDocumentQueryResp
