+++ packages/api/src/external/carequality/outbound-document-retrieval-resp.ts
@@ -1,4 +1,4 @@
-import { OutboundDocumentRetrievalResp as OutboundDocumentRetrievalRespCore } from "@metriport/core/external/carequality/ihe-result";
+import { OutboundDocumentRetrievalResp as OutboundDocumentRetrievalRespCore } from "@metriport/core/external/carequality/ihe-gateway/ihe-result";
 import { BaseDomainCreate } from "@metriport/core/domain/base-domain";
 
 export interface OutboundDocumentRetrievalResp
