+++ packages/core/src/external/carequality/ihe-gateway/ihe-result.ts
